+++ front/pages/w/[wId]/assistant/[cId]/index.tsx
@@ -2,12 +2,13 @@ import type { UserType } from "@dust-tt/types";
 import type { InferGetServerSidePropsType } from "next";
 import { useRouter } from "next/router";
 import type { ReactElement } from "react";
-import { useEffect, useState } from "react";
+import { useCallback, useContext, useEffect, useState } from "react";
 import React from "react";
 
 import { ConversationContainer } from "@app/components/assistant/conversation/ConversationContainer";
 import type { ConversationLayoutProps } from "@app/components/assistant/conversation/ConversationLayout";
 import ConversationLayout from "@app/components/assistant/conversation/ConversationLayout";
+import { InputBarContext } from "@app/components/assistant/conversation/input_bar/InputBarContext";
 import { CONVERSATION_PARENT_SCROLL_DIV_ID } from "@app/components/assistant/conversation/lib";
 import { withDefaultUserAuthRequirements } from "@app/lib/iam/session";
 
@@ -66,6 +67,22 @@ export default function AssistantConversation({
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
   const [conversationKey, setConversationKey] = useState<string | null>(null);
   const router = useRouter();
+  const { animate, setAnimate, setSelectedAssistant } =
+    useContext(InputBarContext);
+
+  const setInputbarMention = useCallback(
+    (agentSid: string) => {
+      setSelectedAssistant({ configurationId: agentSid });
+      setAnimate(true);
+    },
+    [setAnimate, setSelectedAssistant]
+  );
+
+  useEffect(() => {
+    if (animate) {
+      setTimeout(() => setAnimate(false), 500);
+    }
+  });
 
   // This useEffect handles whether to change the key of the ConversationContainer
   // or not. Altering the key forces a re-render of the component. A random number
@@ -91,7 +108,28 @@ export default function AssistantConversation({
         mainTag.scrollTo(0, 0);
       }
     }
-  }, [router.query, setConversationKey, initialConversationId]);
+
+    const handleRouteChange = () => {
+      const assistantId = router.query.assistant ?? null;
+      if (assistantId && typeof assistantId === "string") {
+        setInputbarMention(assistantId);
+      }
+    };
+
+    // Initial check in case the component mounts with the query already set.
+    handleRouteChange();
+
+    router.events.on("routeChangeComplete", handleRouteChange);
+    return () => {
+      router.events.off("routeChangeComplete", handleRouteChange);
+    };
+  }, [
+    router.query,
+    router.events,
+    setConversationKey,
+    initialConversationId,
+    setInputbarMention,
+  ]);
 
   useEffect(() => {
     function handleNewConvoShortcut(event: KeyboardEvent) {
@@ -116,6 +154,8 @@ export default function AssistantConversation({
       owner={owner}
       subscription={subscription}
       user={user}
+      setInputbarMention={setInputbarMention}
+      setAnimate={setAnimate}
     />
   );
 }
